requirements.txt
python-dotenv
anthropic
pytest
pytest-asyncio